+++ src/copick/util/handlers/picks/__init__.py
@@ -4,8 +4,11 @@ This package provides handlers for reading/writing particle picks in various for
 All handlers are automatically registered with the FormatRegistry on import.
 """
 
-from copick.util.handlers import FormatRegistry, csv_handler, dynamo_handler, em_picks_handler
+from copick.util.handlers.picks.csv import csv_handler
+from copick.util.handlers.picks.dynamo import dynamo_handler
+from copick.util.handlers.picks.em import em_picks_handler
 from copick.util.handlers.picks.star import star_handler
+from copick.util.handlers.registry import FormatRegistry
 
 # Register all picks handlers
 FormatRegistry.register_picks_handler(star_handler)
